example update for package g/garray/gmap/gset/ghttp
