+++ container/glist/glist_example_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/gogf/gf/container/glist"
 )
 
-func Example_basic() {
+func ExampleNew() {
 	n := 10
 	l := glist.New()
 	for i := 0; i < n; i++ {
@@ -32,14 +32,14 @@ func Example_basic() {
 	fmt.Println(l.Len())
 
 	// Output:
-	//10
-	//[0 1 2 3 4 5 6 7 8 9]
-	//[9 8 7 6 5 4 3 2 1 0]
-	//0123456789
-	//0
+	// 10
+	// [0 1 2 3 4 5 6 7 8 9]
+	// [9 8 7 6 5 4 3 2 1 0]
+	// 0123456789
+	// 0
 }
 
-func Example_iterate() {
+func ExampleList_RLockFunc() {
 	// concurrent-safe list.
 	l := glist.NewFrom(garray.NewArrayRange(1, 10, 1).Slice(), true)
 	// iterate reading from head.
@@ -63,21 +63,39 @@ func Example_iterate() {
 	})
 
 	fmt.Println()
+	// Output:
+	// 12345678910
+	// 10987654321
+}
 
+func ExampleList_IteratorAsc() {
+	// concurrent-safe list.
+	l := glist.NewFrom(garray.NewArrayRange(1, 10, 1).Slice(), true)
 	// iterate reading from head using IteratorAsc.
 	l.IteratorAsc(func(e *glist.Element) bool {
 		fmt.Print(e.Value)
 		return true
 	})
-	fmt.Println()
+
+	// Output:
+	// 12345678910
+}
+
+func ExampleList_IteratorDesc() {
+	// concurrent-safe list.
+	l := glist.NewFrom(garray.NewArrayRange(1, 10, 1).Slice(), true)
 	// iterate reading from tail using IteratorDesc.
 	l.IteratorDesc(func(e *glist.Element) bool {
 		fmt.Print(e.Value)
 		return true
 	})
+	// Output:
+	// 10987654321
+}
 
-	fmt.Println()
-
+func ExampleList_LockFunc() {
+	// concurrent-safe list.
+	l := glist.NewFrom(garray.NewArrayRange(1, 10, 1).Slice(), true)
 	// iterate writing from head.
 	l.LockFunc(func(list *list.List) {
 		length := list.Len()
@@ -92,21 +110,14 @@ func Example_iterate() {
 	})
 	fmt.Println(l)
 
-	//output:
-	//12345678910
-	//10987654321
-	//12345678910
-	//10987654321
-	//[1,2,3,4,5,M,7,8,9,10]
+	// Output:
+	// [1,2,3,4,5,M,7,8,9,10]
 }
 
-func Example_popItem() {
+func ExampleList_PopBack() {
 	l := glist.NewFrom(g.Slice{1, 2, 3, 4, 5, 6, 7, 8, 9})
 
 	fmt.Println(l.PopBack())
-	fmt.Println(l.PopBacks(2))
-	fmt.Println(l.PopFront())
-	fmt.Println(l.PopFronts(2))
 
 	// Output:
 	// 9
@@ -114,8 +125,34 @@ func Example_popItem() {
 	// 1
 	// [2 3]
 }
+func ExampleList_PopBacks() {
+	l := glist.NewFrom(g.Slice{1, 2, 3, 4, 5, 6, 7, 8, 9})
 
-func Example_join() {
+	fmt.Println(l.PopBacks(2))
+
+	// Output:
+	// [9 8]
+}
+
+func ExampleList_PopFront() {
+	l := glist.NewFrom(g.Slice{1, 2, 3, 4, 5, 6, 7, 8, 9})
+
+	fmt.Println(l.PopFront())
+
+	// Output:
+	// 1
+}
+
+func ExampleList_PopFronts() {
+	l := glist.NewFrom(g.Slice{1, 2, 3, 4, 5, 6, 7, 8, 9})
+
+	fmt.Println(l.PopFronts(2))
+
+	// Output:
+	// [1 2]
+}
+
+func ExampleList_Join() {
 	var l glist.List
 	l.PushBacks(g.Slice{"a", "b", "c", "d"})
