+++ frame/g/g_z_example_test.go
@@ -4,26 +4,19 @@
 // If a copy of the MIT was not distributed with this file,
 // You can obtain one at https://github.com/gogf/gf.
 
-package ghttp_test
+package g_test
 
 import (
 	"github.com/gogf/gf/frame/g"
 	"github.com/gogf/gf/net/ghttp"
 )
 
-func ExampleGetServer() {
+func ExampleServer() {
+	// A hello world example.
 	s := g.Server()
 	s.BindHandler("/", func(r *ghttp.Request) {
 		r.Response.Write("hello world")
 	})
 	s.SetPort(8999)
 	s.Run()
-}
-
-func ExampleClientResponse_RawDump() {
-	response, err := g.Client().Get("https://goframe.org")
-	if err != nil {
-		panic(err)
-	}
-	response.RawDump()
 }
